attempt to use physical motors and tof for API
