+++ lib/pathfinding/simAPI.h
@@ -1,0 +1,39 @@
+/*
+#ifndef API_H
+#define API_H
+#pragma once
+
+#include <string>
+
+class API {
+
+public:
+
+    static int mazeWidth();
+    static int mazeHeight();
+
+    static bool wallFront();
+    static bool wallRight();
+    static bool wallLeft();
+
+    static void moveForward(int distance = 1);
+    static void turnRight();
+    static void turnLeft();
+
+    static void setWall(int x, int y, char direction);
+    static void clearWall(int x, int y, char direction);
+
+    static void setColor(int x, int y, char color);
+    static void clearColor(int x, int y);
+    static void clearAllColor();
+
+    static void setText(int x, int y, const std::string& text);
+    static void clearText(int x, int y);
+    static void clearAllText();
+
+    static bool wasReset();
+    static void ackReset();
+
+};
+#endif
+*/
